Detox_Lung working + WandB and Optuna
This is a commit with a working project folder for the DETOX-Lung project. Further the WandB and Optuna are also working togheter. For Optuna a local DB is created in the results directory under the projectName.

Warning: The config file has been updated and needs to be reviewed from the DETOX_lung config version.