Large cleaning + wandb logging dir
This is the first step to start the cleaning process of the project to arrange all the directories. This will be further updated and improved on the local computer.

Additional part is that the logging results directory is given to the WandB to keep the logging into a logic place