Finished: Repaired cleaning + structuring
This is a commit with the new structure layout to work with the main scripts. The main scripts are positioned in a subdirectory that will get the upper files with the cwd (current working directory) command to retrieve the upper level information.

In this commit repairs and updates have been performed to:
-  Finalise the re-structuring.
- Loss functions adding
- A script for Daniel has also been created.
- Log file with error in result directory