add ViT model
Added the ViT model as another option for the image encoder block (i.e. in place of a CNN).
Unfortunately, the ViT has a lot of hyperparameters relating to the architecture and I am not sure what each one does exactly, so this code (the parameters) may need some cleaning up