make get_config() function update the base_config recursively
up until now, the main keys in the base_config was completely overwritten by the ones in the experiment_config, even if the experiment_config only updated one little parameter in that key. This caused all of the other parameters in the key to be deleted.

Now, the update function is recursive, and only updates a value in the config whenever it is not a dictionary.