rework testMode and add a config file for it
I have added a config file for testMode, which also lets us change other parameters, not just the number of patients to include (e.g. max_epochs, batch_size, etc.).

I also changed how we subsample the dataset in testMode (in the load_dataset_total() function), so that it samples a certain amount of patients in total (train+val+test, while preserving the ratio between each of the three sets)