Augmentation and evaluation update
In this commit is the augmentation split in different sets. This will help to select the parameters used for standard augmentation, however selected augmentation ranges and results need to be printed/plotted for evaluation purposes.

In this commit is also a new dataset used to predict the RP endpoint with a smaller test set and a larger validation set  (0.4,0.4,0.3)