AttentionMaps, CalibrationPlots, loss for TimeEventPrediction
There are quite some updates in this commit. It is all tested on the DETOX-Lung project details.

Attentionmaps can be created and in Integrated Gradients or GradCam in notebook example. Still, need to clean the code more and test the example if still works

Calibration plot can be created and validated with the help of Daniel. This shows the original and refitted calibration plots for a certain given model.

A start has been made to add time-event predictions to the model. This is also needed for the OS study that a event happens at a certain time point